requirements.txt
asgiref==3.5.2
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1.2
django-cloudinary-storage==0.3.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jmespath==1.0.1
Pillow==9.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
